backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
psycopg2-binary
celery
redis
gunicorn
python-dotenv
requests